fix(admin): rename all acc-franchisee to admin
