+++ src/components/layout/Layout.tsx
@@ -3,7 +3,7 @@ import Footer from './footer/Footer';
 import Header from './header/Header';
 import AppRouter from './app-router/AppRouter';
 import { BrowserRouter } from 'react-router-dom';
-import HeaderFranchisee from './header/HeaderFranchisee';
+import HeaderAdmin from './header/HeaderAdmin';
 
 const Layout = (): React.ReactElement => {
   const isAdmin = true;
@@ -11,7 +11,7 @@ const Layout = (): React.ReactElement => {
     <>
       <BrowserRouter basename={process.env.PUBLIC_URL}>
         <div className="container">
-          {isAdmin ? <HeaderFranchisee /> : <Header />}
+          {isAdmin ? <HeaderAdmin /> : <Header />}
           <AppRouter />
         </div>
         <Footer className="home-page-block" />
