+++ src/components/layout/app-router/AppRouter.tsx
@@ -5,7 +5,7 @@ import NotFoundPage from '../../../pages/404';
 import Franchise from '../../../pages/franchise';
 import InProcess from '../../../pages/in-process';
 import Games from '../../../pages/Games';
-import AccountFranchisee from '../../../pages/Account-franchisee';
+import Admin from '../../../pages/Admin';
 
 const AppRouter = (): React.ReactElement => {
   return (
@@ -15,7 +15,7 @@ const AppRouter = (): React.ReactElement => {
       <Route path="/franchise" element={<Franchise />} />
       <Route path="/inprocess" element={<InProcess />} />
       <Route path="/games" element={<Games />} />
-      <Route path="/account-francisee" element={<AccountFranchisee/>}></Route>
+      <Route path="/admin" element={<Admin />}></Route>
     </Routes>
   );
 };
